Use size_t for image width and height

+++ structures/ddpsf.h
@@ -7,7 +7,7 @@
 
 std::vector<schaapcommon::facets::Facet> CreateRectangularPsfs(
     double phase_centre_ra, double phase_centre_dec, double pixel_scale_x,
-    double pixel_scale_y, double trimmed_image_height,
-    double trimmed_image_width, size_t psf_grid_width, size_t psf_grid_height);
+    double pixel_scale_y, size_t trimmed_image_height,
+    size_t trimmed_image_width, size_t psf_grid_width, size_t psf_grid_height);
 
 #endif  // STRUCTURES_DDPSF_H_
